pyproject.toml
[project]
name = "agentic-timetable-planner"
version = "0.1.0"
description = "AI-powered university timetable planner using Google Gemini and Agent Development Kit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "google-generativeai>=0.3.0",
    "google-cloud-aiplatform>=1.38.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "google-cloud-firestore>=2.13.0",
    "aiofiles>=23.2.1",
    "pandas>=2.1.3",
    "orjson>=3.9.10",
    "firebase-admin>=7.1.0",
    "pyjwt>=2.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "models", "services", "config"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
